+++ src/plugins/meteor/assets/prepare-bundle.sh
@@ -5,7 +5,7 @@ set -e
 APP_DIR=/opt/<%= appName %>
 APPNAME=<%= appName %>
 START_SCRIPT=$APP_DIR/config/start.sh
-IMAGE=mup-<%= appName %> 
+IMAGE=mup-<%= appName %>
 
 build_failed() {
   sudo docker start $APPNAME || true
@@ -29,8 +29,10 @@ sudo cat <<EOT > Dockerfile
 FROM <%= dockerImage %>
 RUN mkdir /built_app
 COPY ./ /built_app
+<% for(var key in env) { %>
+ENV <%- key %>=<%- env[key] %>
+<% } %>
 RUN cd  /built_app/programs/server && \
-    <% for(var key in env) { %> <%- key %>=<%- env[key] %> <% } %> && \
     npm install --unsafe-perm
 EOT